rendering-reflections: Have a better render pass implementation to allow to only render terrain or sky or only sky, ... in environments maps, also this allows to not render water itself in environment maps which really does not make sense in WaterTest, yaaaa #wip

+++ src/tdme/utilities/MutableString.cpp
@@ -156,7 +156,7 @@ MutableString& MutableString::insert(int32_t idx, float f, int32_t decimals)
 	// see: http://stackoverflow.com/questions/7123490/how-compiler-is-converting-integer-to-string-and-vice-versa
 	auto integer = static_cast<int>(f);
 	for (auto i = 0; i < decimals; i++) {
-		auto integerDecimal = static_cast<int>(((f - integer) * Math::pow(10.0f, i + 1))) - (10 * static_cast<int>(((f - integer) * Math::pow(10.0f, i))));
+		auto integerDecimal = static_cast<int>(((f - integer) * Math::pow(10.0f, static_cast<float>(i) + 1.0f))) - (10 * static_cast<int>(((f - integer) * Math::pow(10.0f, static_cast<float>(i)))));
 		insert(idx + i, Math::abs(integerDecimal));
 	}
 	insert(idx, '.');
